feat: enhance directory structure output with tree-like ASCII representation
- Updated `generate_structure` function to display directory tree using ASCII symbols for clearer hierarchy
- Added line connectors (├──, └──, │) for better readability
- Implemented dynamic indentation based on directory depth and last item status
- Included ignore list to skip specified files and directories (e.g., .git, .env, .vscode)
- Handled PermissionError gracefully to avoid crashing on restricted directories
