+++ ignore_config.py
@@ -1,0 +1,32 @@
+def get_ignore_files():
+    return {
+        "python": ["__pycache__", ".venv", ".env", ".pytest_cache", ".mypy_cache", ".vscode", ".idea", ".git"],
+        "nextjs": ["node_modules", ".next", "out", ".env", ".vscode", ".idea", ".git"],
+        "nodejs": ["node_modules", ".env", ".vscode", ".idea", ".git"],
+        "express": ["node_modules", ".env", ".vscode", ".idea"],
+        "flutter": ["build", ".dart_tool", ".vscode", ".idea"],
+        "django": ["__pycache__", "venv", ".env", "db.sqlite3", "staticfiles", ".vscode", ".idea"],
+        "react": ["node_modules", ".env", "build", "dist", ".vscode", ".idea"],
+        "angular": ["node_modules", "dist", ".env", ".vscode", ".idea"],
+        "vue": ["node_modules", "dist", ".env", ".vscode", ".idea"],
+        "ruby": [".bundle", "log", "tmp", "vendor/bundle", ".vscode", ".idea"],
+        "java": ["target", ".idea", ".vscode", "bin", "out"],
+        "springboot": ["target", ".idea", ".vscode", "bin", "out"],
+        "php": ["vendor", "node_modules", ".env", ".vscode", ".idea"],
+        "laravel": ["vendor", "node_modules", ".env", ".vscode", ".idea", "storage"],
+        "go": ["vendor", ".vscode", ".idea"],
+        "rust": ["target", ".vscode", ".idea"],
+        "csharp": ["bin", "obj", ".vs", ".vscode", ".idea"],
+        "cpp": ["build", ".vscode", ".idea"],
+        "kotlin": ["build", ".gradle", ".idea", ".vscode"],
+        "swift": ["build", ".vscode", ".idea"],
+        "dotnet": ["bin", "obj", ".vs", ".vscode", ".idea"],
+        "svelte": ["node_modules", "public/build", ".vscode", ".idea"],
+        "electron": ["node_modules", "dist", ".env", ".vscode", ".idea"],
+        "elixir": ["_build", "deps", ".elixir_ls", ".vscode", ".idea"],
+        "haskell": [".stack-work", ".vscode", ".idea"],
+        "terraform": [".terraform", ".terraform.lock.hcl", ".vscode", ".idea"],
+        "rust": ["target", ".vscode", ".idea", "Cargo.lock", "Cargo.toml"],
+        "ansible": [".vscode", ".idea", "*.retry", "inventory/*"],
+        "bash": [".vscode", ".idea", "*.swp", "*.swo", ".DS_Store"],
+    }
